Component updates: Tiler and Ducati MM

+++ packages/titiler/titiler-d2cmap_git.bb
@@ -6,7 +6,7 @@ SECTION = "libs"
 DEPENDS = "titiler-memmgr tisyslink-lib"
 inherit autotools pkgconfig
 
-PR = "r7"
+PR = "r8"
 PV = "0.00+git+${SRCREV}"
 
 SRC_URI = "git://dev.omapzoom.org/pub/scm/tiler/tiler-userspace.git;protocol=git"
@@ -17,10 +17,12 @@ FILES_${PN} += "${libdir}/*.so"
 FILES_${PN}-dev += "${libdir}/*.*a ${libdir}/pkgconfig/"
 FILES_${PN}-dbg += "${libdir}/.debug/"
 
-EXTRA_OECONF += "--enable-tilermgr --enable-tests"
+EXTRA_OECONF += "--enable-tests"
 
 do_compile_prepend() {
-    ln -s ${S}/../utils/testlib.c ${S}/tests/testlib.c
+    if [ ! -e ${S}/tests/testlib.c ]; then
+        ln -s ${S}/../utils/testlib.c ${S}/tests/testlib.c
+    fi
 }
 
 do_stage() {
